refactor: update project structure and configuration files for Smart ARB Translator, including renaming package and adding formatter settings
