+++ .vscode/launch.json
@@ -30,8 +30,7 @@
                 "/Volumes/WD_SN770_2TB/Github/abcx3_flutter/dart_secrets/api_key.txt",
                 "--language_codes",
                 "sv",
-                "--append_lang_code",
-                "0",
+                "--append-lang-code"
                 // "--l10n_directory",
                 // "/Volumes/WD_SN770_2TB/Github/abcx3_flutter/lib/l10n"
             ],
