--- lib/arb_translator.dart
@@ -1,3 +1,0 @@
-void main(List<String> arguments) {
-  print('run command:\n\npub run arb_translator:translate');
-}
