+++ pubspec.yaml
@@ -1,10 +1,15 @@
-name: arb_translator
-description: A command-line utility which helps to translate arb files with the use of google translations.
-version: 3.0.1
-homepage: https://github.com/justkawal/arb_translator.git
+name: smart_arb_translator
+description: An intelligent command-line utility for translating ARB files with Google Translate API, featuring smart change detection and modular architecture.
+version: 1.0.0
+repository: https://github.com/YOUR_USERNAME/smart_arb_translator
+homepage: https://github.com/YOUR_USERNAME/smart_arb_translator
+documentation: https://github.com/YOUR_USERNAME/smart_arb_translator#readme
 
 environment:
   sdk: ">=3.0.1 <4.0.0"
+
+executables:
+  smart_arb_translator: translate
 
 dependencies:
   arb_merge: ^1.2.0
@@ -16,6 +21,7 @@ dependencies:
   path: ^1.9.1
   petitparser: ^6.1.0
   yaml: ^3.1.3
+
 dev_dependencies:
   lints: ^6.0.0
   test: ^1.26.2
